+++ BlackJackGame/BlackJackGame.cpp
@@ -1,9 +1,11 @@
 #include "BlackJackGame.h"
 
 BlackJackGame::BlackJackGame(int numberPlayers,
+                             int numberBots,
                              int initCash,
                              int numberDecks)
     : _countPlayers(numberPlayers)
+    , _countBots(numberBots)
     , _initCash(initCash)
     , _dealer(Dealer(numberDecks)) {}
 
@@ -15,6 +17,10 @@ void BlackJackGame::addPlayers() {
         std::cin >> name;
         _players.push_back(new Player(name, &_dealer, _initCash));
     }
+    for (int i = 0; i < _countBots; i++)
+        _players.push_back(new Player("Bot" + std::to_string(i+1),
+                                         &_dealer,
+                                         _initCash));
 }
 
 void BlackJackGame::connectGame() {
@@ -24,7 +30,7 @@ void BlackJackGame::connectGame() {
 void BlackJackGame::startGame() {
     do {
         std::cout << std::endl;
-        std::deque<Player*> outPlayers = _dealer.playRound();
+        std::deque<IPlayer*> outPlayers = _dealer.playRound();
         for (auto p = _players.begin(); p < _players.end(); p++) {
             if (*p == outPlayers.front()) {
                 delete *p;
