+++ BlackJackGame/BlackJackGame.h
@@ -1,16 +1,20 @@
+#pragma once
+
 #include <deque>
 
-//#include "Dealer.h"
+#include "Dealer.h"
 #include "Player.h"
 
 class BlackJackGame {
 public:
-    std::deque<Player*> _players;
+    std::deque<IPlayer*> _players;
     int _countPlayers;
+    int _countBots;
     int _initCash;
     Dealer _dealer;
 public:
     BlackJackGame(int numberPlayers = 1,
+                  int numberBots = 0,
                   int initCash = 100,
                   int numberDecks = 4);
